introducing CuratorLoadQueuePeon
